debuging error, also + some sanitization and validation

+++ delete.php
@@ -5,11 +5,16 @@ require_once('database.php');
 require_once('head.php');
 $dbh = new databaseaccess;
 $id = $_POST["id"];
-if(!$_SESSION['check']){
-	echo "Please Login";
+$id = filter_var($id, FILTER_SANITIZE_NUMBER_INT);
+if (filter_var($id,FILTER_VALIDATE_INT)&& $id>-1) {
+	if(!$_SESSION['check']){
+		echo "Please Login";
+	}else{
+		$dbh->delete($id);
+		echo "Post Deleted";
+	}
 }else{
-	$dbh->delete($id);
-	echo "Post Deleted";
-}
+	echo "invalid id";
+}	
 require_once('foot.php');
 ?>
